Update: Initial Commit
Add one of my previous dockerfile used my my end to create an
ssh-friendly docker container used to deploy the script remotely
for testing.

Add a directory to contain all header files that will be used in
the script. Add the base file that will be used to interact with
the user.