Update: Add CMake. Define base headers and implement them
Add a cmake file including all the flags that are being used in the
project as of now. Add additional compile-time include directories
to be able to include header files and their implementations, and
build them along with the rest of the project using CMake.

Add a base header file and its implementation. Segregate header
definition and implementation into two separate directories for
better organization.