Update: Add new generic methods. Refactor.
Relatively large-scale modification to the existing generic methods
placed inside `commons` - this commit introduces a new generic method
`raw_gen_str` which intends to replace the existing functionality of
the string generator, while at the same time allowing a lot more
customization depending on the needs of the calling method(s).

Add two new methods to act as an interface to avoid directly interacting
with the existing logic placed inside the new method. Both of these methods
have a pretty similar functioning to the original string generator - that is,
to generate a new string (of the same or greater length), and to populate
it with the data from the source string.