Update: Add a new public method to decrypt Play Fair cipher
Add a new public method that can be used to decrypt text from the
existing cipher play-fair cipher method.

The internal functioning of this method is very much similar to the
working of the original play fair cipher (one that encrypts strings).

This commit also includes a (minor) bug-patch. Instead of directly
padding the source message string when required, the program will
now start by generating a copy of the source, and then making
modifications to this copy as needed. Reduces the possibility
of an overflow during run-time.