Minor: Small scale refactor
Small scale refactoring of the code base to maintain a logical
order of the methods/objects.

Move the methods to validate and extract data from a string using
regex pattern with the rest of the helpers. This way, the helper
methods can be used by the rest of the program and the logical
consistency of the code-base is maintained.