Update: Modify main method
Modify the source root file including the main method to bring the
project up-to-date with the changes made in the previous commit(s).

This commit removes a large part of the logic previously present in
the main method - including the part dealing with command-line
arguments. Most of this logic has already been delegated to a helper
method in the previous commit that will handle all of this internally.

This commit makes the main method be tidy (again) - involving only
the minimal logic required in the main method, and delegating the
important parts of logic to structured code hidden behind method
calls and/or private interfaces that are present in individual source
files (without revealing them to the rest of the project).